README.md
Simple Web Application using Spring boot, hibernate, thymeleaf and postgreSQL Database

In this project, Java Spring boot has been used as a backend which consists of inbuilt tomcat server. For the front end thymeleaf has been used. When the tomcat server is started, the web pages can be accessed in the browser. After running the program, the login page will appear.

First, the user needs to be registered as USER Role. So, after submitting the registration page, the user will receive an email stating that the account has be created. Then the user can deposit the amount, and only then user can withdraw and transfer the amount. The user can also view the statement as well as update the additional details.

The ADMIN Role can do all the above-mentioned activities and also the admin can view all the details of the account holders. For security purpose, Spring security has been used.

For this project, Java Spring boot, thymeleaf, hibernate, spring security, PostgreSQL Database, CSS and javascript is used.